UniProtParser: confirm source IDs have been found
To make the check in run() simpler, source IDs are now stored in a hash
rather than a bunch of scalars; that way all we need is a single List::Util::any
test to check all sources at this level. In the near future it will also
simplify the passing of source IDs between functions.