UniProtParser: only print release information once done with the release file
The way the code is presently structured it is theoretically possible for the
release file to contain multiple matching lines for either source, with each
subsequent appearance overwriting the previously extracted respective value.
Given it is the final extracted values that are used for anything, I would say
there is no point in printing out the discarded ones.